Fix tests

+++ src/routers/xcm/Router.sol
@@ -7,7 +7,7 @@ import {ConnectorMessages} from "../../Messages.sol";
 
 interface ConnectorLike {
   function addPool(uint64 poolId) external;
-  function addTranche(uint64 poolId, bytes16 trancheId, string memory tokenName, string memory tokenSymbol, uint256 price) external;
+  function addTranche(uint64 poolId, bytes16 trancheId, string memory tokenName, string memory tokenSymbol, uint128 price) external;
   function updateMember(uint64 poolId, bytes16 trancheId, address user, uint64 validUntil) external;
   function updateTokenPrice(uint64 poolId, bytes16 trancheId, uint256 price) external;
   function handleTransfer(uint64 poolId, bytes16 trancheId, address user, uint256 amount) external;
@@ -47,7 +47,7 @@ contract ConnectorXCMRouter {
             uint64 poolId = ConnectorMessages.parseAddPool(_msg);
             connector.addPool(poolId);
         } else if (ConnectorMessages.isAddTranche(_msg)) {
-            (uint64 poolId, bytes16 trancheId, string memory tokenName, string memory tokenSymbol, uint256 price) = ConnectorMessages.parseAddTranche(_msg);
+            (uint64 poolId, bytes16 trancheId, string memory tokenName, string memory tokenSymbol, uint128 price) = ConnectorMessages.parseAddTranche(_msg);
             connector.addTranche(poolId, trancheId, tokenName, tokenSymbol, price);
         } else if (ConnectorMessages.isUpdateMember(_msg)) {
             (uint64 poolId, bytes16 trancheId, address user, uint64 validUntil) = ConnectorMessages.parseUpdateMember(_msg);
